Implement CI using Github Actions
This change implements a continuous integration (CI) workflow using
Github Actions.

In particular, we configure workflows to run unit and end-to-end tests
on every PR (update) and push to the master branch. This includes the
necessary steps to build test runner and plugin images and execute
upstream storage tests against a dynamically created DOKS cluster.

We also trigger the release process when a tag is pushed, including the
publication of an official Docker Hub container image as well as the
creation of a Github release.

This change obsoletes our usage of Travis CI.